requirements.txt
git+https://github.com/BBottoml/python-chromedriver-binary.git
cycler==0.10.0
fpdf2==2.0.3
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.16.4
pandas==0.25.0
Pillow==6.1.0
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
selenium==3.141.0
six==1.12.0
urllib3==1.25.3
